Improving text alignment.
For good results, text needs to use different extents for alignment than the actual extents. For example, when left-aligning the digit '1', it may need to be shifted a bit to the right, to properly align with other digits.

This change introduces 'anchorExtents()' method to the Drawable interface. The default implementation simply returns extents(). The text label overrides this method to return the extents based on the font and glyph metrics. Because of this, left- and right-alignment will use the origin and advance (rather than the actual pixel extents), and the top and bottom alignment will use font ascent and descent. (Use tNoVAlign to keep the text anchored at the baseline).

A side-effect of this change is that text can sometimes get clipped, because the pixels can actually extend beyond the 'anchorExtents'. Generally, when drawing text in bounded containers, you should use some small constant padding to overcome this.