Seriously revamping alignment.
Making it possible to specify two anchors, e.g. align the baseline of text to the center of the screen (with an optional offset). Also, making it possible to capture sub-regions of devices and align content within them.

This functionality makes it easier to avoid pixel-level tuning, and implement good-looking GUIs that look good on different screens (with different resolutions).