Fixes in the GT911 touch driver, and the Matouch combo that uses it.
1. A number of spurious errors reported at startup;
2. Occasional restarts quickly after startup.

One of the fixes is to hold down the RESET pin for much longer than the spec says (which is only 100 microseconds). In practice, for Matouch devices, it needs to be held down for as much as 300 milliseconds!

To make this delay less impactful for the device start time, reset in this driver is now asynchronous - the touch simply doesn't respond until reset completes, but otherwise the device isn't blocked on it.